# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Relay API
  version: 0.1.0
  summary: API-first second-brain backend for long-running AI-assisted work.
servers:
  - url: http://127.0.0.1:8080
    description: Local development
  - url: https://relay.4gly.dev
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque token
  schemas:
    SuccessEnvelope:
      type: object
      required: [ok, command, data, warnings]
      properties:
        ok:
          type: boolean
          const: true
        command:
          type: string
        data: {}
        warnings:
          type: array
          items:
            type: string
    ErrorEnvelope:
      type: object
      required: [ok, command, error]
      properties:
        ok:
          type: boolean
          const: false
        command:
          type: string
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code:
              type: string
            message:
              type: string
            retryable:
              type: boolean
            missing_fields:
              type: array
              items:
                type: string
...
    StepStatus:
      type: string
      enum: [ok, failed, skipped]
      description: |
        Locked validation chip status (D2 + E4):
        - ok: this step succeeded
        - failed: this step is the cause of the error
        - skipped: an earlier step's failure makes this step undefined (we did not probe further)
    ValidationStep:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
          description: |
            Stable chip identifier. The server emits exactly two chips per probe:
            anthropic_key (the API rejected/accepted the key) and
            anthropic_quota (the account has remaining quota).
          enum: [anthropic_key, anthropic_quota]
        status:
          $ref: '#/components/schemas/StepStatus'
    OnboardingCompleteInput:
      type: object
      required: [anthropic_key]
      properties:
        anthropic_key:
          type: string
          description: |
            The user's Anthropic API key (must start with sk-ant-). Encrypted at rest. Transmitted only to Anthropic for validation, never stored or logged in plaintext. The strict decoder rejects unknown fields (including the legacy relay_url field that D3 removed) with code UNKNOWN_JSON_FIELD.
    OnboardingCompleteResult:
      type: object
      required: [onboarding_complete, default_project_id, anthropic_key_prefix, anthropic_key_last4, steps]
      properties:
        onboarding_complete:
          type: boolean
        default_project_id:
          type: string
          description: ID of the user's "Personal" project, created idempotently on first onboarding (D4).
        anthropic_key_prefix:
          type: string
          description: First 18 characters of the key followed by an ellipsis. Safe to display in settings UI.
        anthropic_key_last4:
          type: string
          description: Last 4 characters of the key. Safe to display in settings UI.
        steps:
          type: array
          description: Always two entries (anthropic_key, anthropic_quota) regardless of success or failure (D2). Frame 3 chips render directly from this array.
          items:
            $ref: '#/components/schemas/ValidationStep'
    OnboardingCompleteResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OnboardingCompleteResult'
    OnboardingDeleteData:
      type: object
      required: [status]
      properties:
        status:
          type: string
          const: ok
    OnboardingDeleteResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OnboardingDeleteData'
...
paths:
...
  /v1/auth/{provider}/start:
    get:
      tags: [auth]
      summary: Begin an OAuth sign-in by redirecting to the provider
      operationId: startOAuth
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [github, google]
        - name: redirect_to
          in: query
          required: false
          description: Optional in-app path to land on after callback completes.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to provider authorization URL
        '404':
          description: Provider is not configured on this Relay deployment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/auth/{provider}/callback:
    get:
      tags: [auth]
      summary: OAuth callback that issues a session cookie and redirects back
      operationId: oauthCallback
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [github, google]
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the original redirect_to (or `/` when empty)
        '400':
          description: Missing required code or state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: OAuth state invalid, expired, or already consumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Provider is not configured on this Relay deployment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: Upstream OAuth provider exchange failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /v1/onboarding:
    post:
      tags: [onboarding]
      summary: Validate and store the user's Anthropic API key
      description: |
        Validates the supplied Anthropic API key against api.anthropic.com /v1/models
        and, on success, envelope-encrypts it under the active KEK and persists it for
        the cookie-session user. Idempotent: re-onboarding rotates the ciphertext and
        the row's aad_salt while preserving default_project_id.

        Trust promise: encrypted at rest. Transmitted only to Anthropic for validation,
        never stored or logged in plaintext.

        Auth: cookie session. Admin bearer is rejected (no user identity to bind the
        row to). Strict decoder: any unknown field (including the legacy relay_url
        field removed in D3) returns 400 UNKNOWN_JSON_FIELD.
      operationId: onboardingComplete
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardingCompleteInput'
      responses:
        '200':
          description: Onboarding completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingCompleteResponse'
        '400':
          description: |
            Validation failed. error.code is one of:
            MISSING_REQUIRED_FIELDS, UNKNOWN_JSON_FIELD,
            INVALID_ANTHROPIC_KEY, ANTHROPIC_QUOTA.
            steps[] is always populated so Frame 3 can render chips.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing session cookie or admin-only bearer (admin has no user identity to bind to)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Server misconfiguration (e.g. RELAY_DATA_ENCRYPTION_KEY missing)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: Could not reach api.anthropic.com (ANTHROPIC_UNREACHABLE)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [onboarding]
      summary: Forget the user's Anthropic API key (D5 trust escape hatch)
      description: |
        NULLs the encrypted key + nonce + aad_salt + onboarding_completed_at.
        The row remains so that prior packets bound to default_project_id still
        resolve, and so re-onboarding can rotate to a fresh ciphertext +
        aad_salt without losing project history.
      operationId: onboardingDeleteKey
      responses:
        '200':
          description: Key forgotten
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingDeleteResponse'
        '401':
          description: Missing session cookie or admin-only bearer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: User has not completed onboarding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'